pyproject.toml
[project]
name = "agent-fuzzing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "angr==9.2.170",
    "annotated-types==0.7.0",
    "antlr4-python3-runtime==4.13.2",
    "anyio==4.10.0",
    "archinfo==9.2.170",
    "asciimatics==1.15.0",
    "bitarray==3.6.1",
    "bitstring==4.3.1",
    "cachetools==6.1.0",
    "capstone==5.0.3",
    "cart==1.2.3",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "claripy==9.2.170",
    "cle==9.2.170",
    "click==8.1.7",
    "contourpy==1.3.3",
    "cxxheaderparser==1.5.4",
    "cycler==0.12.1",
    "dacite==1.8.1",
    "defusedxml==0.7.1",
    "dill==0.4.0",
    "distro==1.9.0",
    "enum-compat==0.0.3",
    "first==2.0.2",
    "fonttools==4.59.0",
    "gevent==25.5.1",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "jiter==0.10.0",
    "jsonpath-ng==1.6.1",
    "keystone-engine==0.9.2",
    "kiwisolver==1.4.9",
    "loguru==0.7.2",
    "markdown-it-py==4.0.0",
    "matplotlib==3.10.5",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "mulpyplexer==0.9",
    "multiprocess==0.70.18",
    "networkx==3.5",
    "numpy==2.3.2",
    "openai==1.102.0",
    "overrides==7.7.0",
    "packaging==25.0",
    "pefile==2024.8.26",
    "pillow==10.4.0",
    "ply==3.11",
    "prompt-toolkit==3.0.51",
    "protobuf==6.31.1",
    "psutil==7.0.0",
    "pycparser==2.22",
    "pycryptodome==3.23.0",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydemumble==0.0.1",
    "pydot==4.0.1",
    "pyelftools==0.32",
    "pyfiglet==1.0.2",
    "pyformlang==1.0.11",
    "pygments==2.19.2",
    "pyparsing==3.2.3",
    "pypcode==3.3.0",
    "pyperclip==1.9.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-fx==0.3.2",
    "python-registry==1.3.1",
    "pyvex==9.2.170",
    "pyyaml==6.0.2",
    "qiling==1.4.6",
    "questionary==2.1.0",
    "rich==14.1.0",
    "setuptools==80.9.0",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "sympy==1.14.0",
    "termcolor==3.1.0",
    "tqdm==4.67.1",
    "typing-extensions==4.12.2",
    "typing-inspection==0.4.1",
    "unicodecsv==0.14.1",
    "unicorn==2.1.3",
    "unique-log-filter==0.1.0",
    "urwid==2.6.15",
    "wcwidth==0.2.13",
    "yamale==5.2.1",
    "z3-solver==4.13.0.0",
    "zope-event==5.1.1",
    "zope-interface==7.2",
]
